Move HTTPClientStub to a shared scope in a separate file
