+++ EssentialApp/EssentialAppTests/Helpers/HttpClientStub.swift
@@ -1,0 +1,36 @@
+//
+//  HttpClientStub.swift
+//  EssentialAppTests
+//
+//  Created by ULLAS JAIN on 31/10/23.
+//
+
+import EssentialFeed
+import Foundation
+
+class HTTPClientStub: HTTPClient {
+    private class Task: HTTPClientTask {
+        func cancel() {}
+    }
+
+    private let stub: (URL) -> HTTPClient.Result
+
+    init(stub: @escaping (URL) -> HTTPClient.Result) {
+        self.stub = stub
+    }
+
+    func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
+        completion(stub(url))
+        return Task()
+    }
+}
+
+extension HTTPClientStub {
+    static var offline: HTTPClientStub {
+        HTTPClientStub(stub: { _ in .failure(NSError(domain: "offline", code: 0)) })
+    }
+
+    static func online(_ stub: @escaping (URL) -> (Data, HTTPURLResponse)) -> HTTPClientStub {
+        HTTPClientStub { url in .success(stub(url)) }
+    }
+}
